requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-geojson==3.2.0
django-jazzmin==2.4.8
django-leaflet==0.28.1
django-material==1.9.0
gunicorn==20.1.0
idna==3.2
oauthlib==3.1.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.2.0
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
